remove final ammonite dependencies
