+++ main/util/src/mill/util/Loggers.scala
@@ -97,6 +97,45 @@ case class PrefixLogger(out: ColorLogger, context: String, tickerContext: String
   override def debugEnabled: Boolean = out.debugEnabled
 }
 
+
+class AnsiNav(output: Writer){
+  def control(n: Int, c: Char) = output.write("\u001b[" + n + c)
+
+  /**
+   * Move up `n` squares
+   */
+  def up(n: Int) = if (n == 0) "" else control(n, 'A')
+  /**
+   * Move down `n` squares
+   */
+  def down(n: Int) = if (n == 0) "" else control(n, 'B')
+  /**
+   * Move right `n` squares
+   */
+  def right(n: Int) = if (n == 0) "" else control(n, 'C')
+  /**
+   * Move left `n` squares
+   */
+  def left(n: Int) = if (n == 0) "" else control(n, 'D')
+
+  /**
+   * Clear the screen
+   *
+   * n=0: clear from cursor to end of screen
+   * n=1: clear from cursor to start of screen
+   * n=2: clear entire screen
+   */
+  def clearScreen(n: Int) = control(n, 'J')
+  /**
+   * Clear the current line
+   *
+   * n=0: clear from cursor to end of line
+   * n=1: clear from cursor to start of line
+   * n=2: clear entire line
+   */
+  def clearLine(n: Int) = control(n, 'K')
+}
+
 case class PrintLogger(
     override val colored: Boolean,
     disableTicker: Boolean,
@@ -140,7 +179,7 @@ case class PrintLogger(
         case PrintState.Ticker =>
           val p = new PrintWriter(infoStream)
           // Need to make this more "atomic"
-          val nav = new ammonite.terminal.AnsiNav(p)
+          val nav = new AnsiNav(p)
           nav.up(1)
           nav.clearLine(2)
           nav.left(9999)
